♻️ Refactor: 로그인 관련 type 폴더 정리
