+++ src/pages/login/hooks/useKakaoLogin.ts
@@ -1,29 +1,16 @@
 import { useMutation } from '@/hooks/useApi';
 import useNavigation from '@/hooks/useNavigation';
+import type {
+  KakaoLoginRequest,
+  KakaoLoginResult,
+} from '@/types/api/login.type';
+import type { ServerApiResponse } from '@/types/api/server.type';
 import { tokenManager } from '@/utils/apiClient';
-
-interface KakaoLoginRequest {
-  code: string;
-  redirectUri: string;
-}
-
-interface ApiResponse<T> {
-  isSuccess: boolean;
-  code: string;
-  message: string;
-  result: T;
-}
-
-interface KakaoLoginResult {
-  userId: number;
-  accessToken: string;
-  refreshToken: string;
-}
 
 export const useKakaoLogin = () => {
   const { navigateTo } = useNavigation();
 
-  return useMutation<ApiResponse<KakaoLoginResult>, KakaoLoginRequest>(
+  return useMutation<ServerApiResponse<KakaoLoginResult>, KakaoLoginRequest>(
     (data) => ({
       method: 'POST',
       url: '/auth/kakao',
@@ -32,7 +19,7 @@ export const useKakaoLogin = () => {
     }),
     {
       onSuccess: (response) => {
-        const data = response as ApiResponse<KakaoLoginResult>;
+        const data = response as ServerApiResponse<KakaoLoginResult>;
 
         if (!data.isSuccess) {
           console.error(data.message);
@@ -45,7 +32,7 @@ export const useKakaoLogin = () => {
         tokenManager.setRefreshToken(refreshToken);
         localStorage.setItem('userId', String(userId));
 
-        console.log('카카오 로그인 성공, userId:', userId);
+        // console.log('카카오 로그인 성공, userId:', userId);
         navigateTo('/my-role');
       },
